# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "version": "1.0.0",
    "title": "Node.js API model",
    "description": "Node.js Express API Microservice with Mongo",
    "contact": {
      "email": "david.a.m@live.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8080/api"
    },
    {
      "url": "https://localhost:8080/api"
    }
  ],
...
  "paths": {
...
    "/products": {
      "get": {
        "tags": ["Products"],
        "summary": "Get all products",
        "description": "Get all products",
        "operationId": "getProducts",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid token, user unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "No products found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "token_auth": []
          }
        ]
      },
      "post": {
        "tags": ["Products"],
        "summary": "Add a new product to the store",
        "description": "Add a new product to the store",
        "operationId": "postProducts",
        "requestBody": {
          "description": "Create a new product in the store",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewProduct"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Invalid token, user unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "token_auth": []
          }
        ]
      }
    },
    "/products/{id}": {
      "get": {
        "tags": ["Products"],
        "summary": "Get product by ID",
        "description": "Find a product searching by ID",
        "operationId": "getProductById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The MongoID of the product to be fetched",
            "required": true,
            "explode": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID value",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Invalid token, user unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "token_auth": []
          }
        ]
      },
      "delete": {
        "tags": ["Products"],
        "summary": "Delete an existing product",
        "description": "Delete an existing product by Id",
        "operationId": "deleteProducts",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The MongoID of the product to be deleted",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Invalid product ID supplied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Invalid token, user unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Product ID not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "token_auth": []
          }
        ]
      },
      "put": {
        "tags": ["Products"],
        "summary": "Update an existing product",
        "description": "Update an existing product by Id",
        "operationId": "updateProducts",
        "requestBody": {
          "description": "Update an existent product in the store",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewProduct"
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The MongoID of the product to be fetched",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Invalid MongoID supplied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Invalid token, user unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "There is no product with that MongoID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "token_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Category": {
        "required": ["name", "category"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 24,
            "maxLength": 24,
            "example": "6208136134ffbeeabc4b1ead"
          },
          "name": {
            "type": "string",
            "example": "PASTRIES"
          },
          "status": {
            "type": "boolean",
            "description": "current status in the store"
          }
        }
      },
...
      "NewProduct": {
        "required": ["name", "category"],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "PASTRY ROLL"
          },
          "categoryId": {
            "type": "string",
            "minLength": 24,
            "maxLength": 24,
            "example": "62083fbdb5d9510f71cf6988"
          },
          "price": {
            "type": "numeric",
            "example": 10.0
          },
          "description": {
            "type": "string",
            "example": "A delicious pastry roll"
          },
          "available": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
      "Product": {
        "required": ["name", "category"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 24,
            "maxLength": 24,
            "example": "68376272814ffbeeabc4b1ed"
          },
          "name": {
            "type": "string",
            "example": "PASTRY ROLL"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "status": {
            "type": "boolean",
            "description": "current status in the store"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "BadRequest"
          },
          "title": {
            "type": "string",
            "example": "Invalid ID"
          },
          "detail": {
            "type": "string",
            "example": "The ID supplied is not a valid MongoID value"
          }
        }
      }
    },
    "securitySchemes": {
      "token_auth": {
        "type": "apiKey",
        "name": "x-token",
        "in": "header"
      }
    }
  }
}